remove unneeded cast
git-svn-id: https://svn.apache.org/repos/asf/lucene/java/trunk@834250 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/org/apache/lucene/util/TestPriorityQueue.java
@@ -61,7 +61,7 @@ public class TestPriorityQueue extends LuceneTestCase {
         int last = Integer.MIN_VALUE;
         for (int i = 0; i < count; i++)
         {
-            Integer next = (Integer)pq.pop();
+            Integer next = pq.pop();
             assertTrue(next.intValue() >= last);
             last = next.intValue();
             sum2 += last;
